Add js_of_ocaml stub with overlay demo
This seemed like the path of least resistance to get an interactive interface,
both in terms of implementing it in and/or interfacing it with the ocaml
and in terms of sharing the result.

The included demo is pretty bare-bones, but I think still useful to
explore corner cases quickly.

I renamed the main dwarf-locstack.ml file because the ocaml bytecode
compiler was complaining that it didn't have a corresponding module
identifier. I don't have a ton of experience with ocaml so if there's a
better way to address this I can undo that change.