DW_OP_call, DW_OP_implicit_pointer: refer to locexpr with a name
Instead of embedding the location expression that is referred in
DW_OP_call and DW_OP_implicit_pointer within the operator, use a name
and find the location expression associated with that name in the
context.  Use DW_AT_location in the context for specifying the
location expressions of other DIEs.

Signed-off-by: Tankut Baris Aktemur <tankut.baris.aktemur@intel.com>